feat: add pylint C0302 (too-many-lines) check to complement Ruff linting
- Add pylint to dev dependencies with version constraint >=3.3.7
- Configure pylint in pyproject.toml to only enable C0302 check with 400-line limit
- Add pylint pre-commit hook for src directory only
- Fix minor ruff S110 warning in extractor.py
