+++ src/oboyu/common/onnx/__init__.py
@@ -1,0 +1,24 @@
+"""ONNX conversion and optimization utilities."""
+
+from .cross_encoder_model import (
+    ONNXCrossEncoderModel,
+    convert_cross_encoder_to_onnx,
+    get_or_convert_cross_encoder_onnx_model,
+)
+from .embedding_model import (
+    ONNXEmbeddingModel,
+    convert_to_onnx,
+    get_or_convert_onnx_model,
+)
+from .quantization import QUANTIZATION_AVAILABLE, quantize_model_dynamic
+
+__all__ = [
+    "QUANTIZATION_AVAILABLE",
+    "quantize_model_dynamic",
+    "ONNXEmbeddingModel",
+    "convert_to_onnx",
+    "get_or_convert_onnx_model",
+    "ONNXCrossEncoderModel",
+    "convert_cross_encoder_to_onnx",
+    "get_or_convert_cross_encoder_onnx_model",
+]
