+++ src/oboyu/indexer/models/tokenizer_service.py
@@ -20,110 +20,10 @@ try:
 except ImportError:
     HAS_JAPANESE_TOKENIZER = False
 
+# Import Japanese stop words from separate module
+from .stop_words import DEFAULT_JAPANESE_STOP_WORDS
+
 logger = logging.getLogger(__name__)
-
-
-# Default Japanese stop words
-DEFAULT_JAPANESE_STOP_WORDS = {
-    # Particles
-    "は",
-    "が",
-    "を",
-    "に",
-    "で",
-    "と",
-    "も",
-    "や",
-    "の",
-    "へ",
-    "から",
-    "まで",
-    "より",
-    "ね",
-    "よ",
-    # Auxiliary verbs
-    "です",
-    "ます",
-    "だ",
-    "である",
-    "でした",
-    "ました",
-    "でしょう",
-    "ましょう",
-    # Pronouns
-    "これ",
-    "それ",
-    "あれ",
-    "どれ",
-    "この",
-    "その",
-    "あの",
-    "どの",
-    # Common words
-    "こと",
-    "もの",
-    "ため",
-    "とき",
-    "ところ",
-    "ほう",
-    "さん",
-    "くん",
-    "ちゃん",
-    # English common words in Japanese text
-    "the",
-    "a",
-    "an",
-    "and",
-    "or",
-    "but",
-    "in",
-    "on",
-    "at",
-    "to",
-    "for",
-    "of",
-    "with",
-    # High-frequency verbs
-    "為る",
-    "する",
-    "し",
-    "居る",
-    "いる",
-    "い",
-    "成る",
-    "なる",
-    "なり",
-    "なっ",
-    "有る",
-    "ある",
-    "あり",
-    "言う",
-    "いう",
-    "思う",
-    "見る",
-    "出る",
-    "来る",
-    "行く",
-    "分かる",
-    "わかる",
-    "知る",
-    "聞く",
-    "読む",
-    "書く",
-    "話す",
-    "働く",
-    "立つ",
-    "座る",
-    "歩く",
-    "走る",
-    "食べる",
-    "飲む",
-    "寝る",
-    "起きる",
-    "買う",
-    "売る",
-    "作る",
-}
 
 # POS tags to exclude from indexing (too generic or not content-bearing)
 EXCLUDED_POS_TAGS = {
