+++ src/oboyu/indexer/storage/utils.py
@@ -1,0 +1,14 @@
+"""Utility classes and functions for storage operations."""
+
+import json
+from datetime import datetime
+
+
+class DateTimeEncoder(json.JSONEncoder):
+    """JSON encoder that handles datetime objects."""
+
+    def default(self, obj: object) -> object:
+        """Convert datetime objects to ISO format strings."""
+        if isinstance(obj, datetime):
+            return obj.isoformat()
+        return super().default(obj)
